# server/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"PingResponse": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"email",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Language": {
				"properties": {
					"code": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"native": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"name",
					"native"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CARDTYPES.WHITE": {
				"enum": [
					"white"
				],
				"type": "string"
			},
			"CARDTYPES.BLACK": {
				"enum": [
					"black"
				],
				"type": "string"
			},
			"CardType": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/CARDTYPES.WHITE"
					},
					{
						"$ref": "#/components/schemas/CARDTYPES.BLACK"
					}
				]
			},
			"Cardpack": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"nsfw": {
						"type": "boolean"
					},
					"userId": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"languageCode": {
						"type": "string"
					},
					"language": {
						"$ref": "#/components/schemas/Language"
					},
					"blackCards": {
						"items": {
							"$ref": "#/components/schemas/BlackCard"
						},
						"type": "array"
					},
					"whiteCards": {
						"items": {
							"$ref": "#/components/schemas/WhiteCard"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name",
					"description",
					"nsfw",
					"userId",
					"user",
					"languageCode",
					"language",
					"blackCards",
					"whiteCards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BlackCard": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"$ref": "#/components/schemas/CardType"
					},
					"text": {
						"type": "string"
					},
					"cardpackId": {
						"type": "string"
					},
					"cardpack": {
						"$ref": "#/components/schemas/Cardpack"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"type",
					"text",
					"cardpackId",
					"cardpack"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WhiteCard": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"$ref": "#/components/schemas/CardType"
					},
					"text": {
						"type": "string"
					},
					"cardpackId": {
						"type": "string"
					},
					"cardpack": {
						"$ref": "#/components/schemas/Cardpack"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"type",
					"text",
					"cardpackId",
					"cardpack"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICardpackPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"nsfw": {
						"type": "boolean"
					},
					"userId": {
						"type": "string"
					},
					"languageCode": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"nsfw",
					"userId",
					"languageCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateCardpackPayload": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"nsfw": {
						"type": "boolean"
					},
					"userId": {
						"type": "string"
					},
					"languageCode": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"nsfw",
					"userId",
					"languageCode",
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Card": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"$ref": "#/components/schemas/CardType"
					},
					"text": {
						"type": "string"
					},
					"cardpackId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"type",
					"text",
					"cardpackId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICardPayload": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/CardType"
					},
					"text": {
						"type": "string"
					},
					"cardpackId": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"text",
					"cardpackId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CardsOfCardpack": {
				"properties": {
					"whiteCards": {
						"items": {
							"$ref": "#/components/schemas/WhiteCard"
						},
						"type": "array"
					},
					"blackCards": {
						"items": {
							"$ref": "#/components/schemas/BlackCard"
						},
						"type": "array"
					}
				},
				"required": [
					"whiteCards",
					"blackCards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserResponse": {
				"properties": {
					"user": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"accessToken": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"user",
					"accessToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ILoginPayload": {
				"properties": {
					"tokenId": {
						"type": "string"
					}
				},
				"required": [
					"tokenId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRegisterPayload": {
				"properties": {
					"tokenId": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"tokenId",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IValidateResponse": {
				"properties": {
					"valid": {
						"type": "boolean"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"valid",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IValidateUsernamePayload": {
				"properties": {
					"username": {
						"type": "string"
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IValidateTokenIdPayload": {
				"properties": {
					"tokenId": {
						"type": "string"
					}
				},
				"required": [
					"tokenId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "card-pack-manager",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"email": "maijaaaw@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/ping": {
			"get": {
				"operationId": "GetMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PingResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Ping"
				],
				"security": [],
				"parameters": []
			}
		},
		"/cardpacks": {
			"get": {
				"operationId": "GetCardpacks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Cardpack"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Cardpack"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateCardpack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Cardpack"
								}
							}
						}
					}
				},
				"tags": [
					"Cardpack"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICardpackPayload"
							}
						}
					}
				}
			}
		},
		"/cardpacks/id": {
			"post": {
				"operationId": "UpdateCardpack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cardpack"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cardpack"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateCardpackPayload"
							}
						}
					}
				}
			}
		},
		"/cardpacks/{id}": {
			"get": {
				"operationId": "GetCardpack",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Cardpack"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cardpack"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards": {
			"post": {
				"operationId": "CreateCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Card"
								}
							}
						}
					}
				},
				"tags": [
					"Card"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICardPayload"
							}
						}
					}
				}
			}
		},
		"/cards/{id}": {
			"get": {
				"operationId": "GetCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Card"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Card"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards/cardpack/{id}": {
			"get": {
				"operationId": "GetCardsByCardpackId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CardsOfCardpack"
								}
							}
						}
					}
				},
				"tags": [
					"Card"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ILoginPayload"
							}
						}
					}
				}
			}
		},
		"/users/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRegisterPayload"
							}
						}
					}
				}
			}
		},
		"/users/register/validate/username": {
			"post": {
				"operationId": "ValidateUsername",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IValidateUsernamePayload"
							}
						}
					}
				}
			}
		},
		"/users/register/validate/tokenId": {
			"post": {
				"operationId": "ValidateTokenId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IValidateResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IValidateTokenIdPayload"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}